+++ api/answers.py
@@ -23,7 +23,7 @@ class Settings(BaseSettings):
 
 
 class Answer(BaseModel):
-    day: Optional[str] = datetime.date.today().strftime("%Y-%m-%d")
+    day: int = int(datetime.date.today().strftime("%Y%m%d"))
     word: str
 
 
@@ -53,7 +53,7 @@ def get_logger():
 
 # get WOTD based on the date parameter entered
 @app.get("/answers/{day}")
-def get_answer(day: str, response: Response, db: sqlite3.Connection = Depends(get_db)):
+def get_answer(day: int, response: Response, db: sqlite3.Connection = Depends(get_db)):
     cur = db.execute("SELECT word FROM answers WHERE day = ?", [day])
     wotd = cur.fetchall()
     if not wotd:
@@ -95,7 +95,7 @@ def change_answer(
 @app.get("/check/{guess}", response_model=Check)
 def find_answer(
     guess: str,
-    day: str = datetime.date.today().strftime("%Y-%m-%d"),
+    day: int = int(datetime.date.today().strftime("%Y%m%d")),
     db: sqlite3.Connection = Depends(get_db),
     logger: logging.Logger = Depends(get_logger),
 ):